Change IN-SUITE* to update the suite.
See also issue #1.

With this change, the properties, including :IN, are updated while
keeping the suite object alive.  This means however, that because of the
default value of *SUITE* for the parent, resetting the relationship is
error-prone, so an explicit self-check is inserted to prevent a link
from a suite onto itself.

I would be really happy to change the default to T instead of *SUITE*.
That would also make things more explicit in the long run.

Included is also a testcase, which would previously fail (to update the
old object).